fix: sync status after runner crash
We were not updating the sync status when a sync was aborted.
It used to work for manual cancellation because in this case the runner
is actually returning a failed task but when a task is expired after a
runner crash, there is no failed task returned by the runner.
In all case we should update the syncJob status when a sync is aborted
One day I will tackle a refactoring of the syncJob table so we don't
have to keep the status in-sync, but for now I am just fixing the bug.