feat(rate-limit): separate rate limit buckets for scripts and direct calls
Add isScript flag to node client and use it to differentiate rate limiting
buckets between script executions and regular customers API calls.
The objective is to ensure direct requests to the API are not consuming
the same rate-limit as script and disturbed each other. When it happens
it is confusing to customers where the requests come from since some
requests from the scripts are done by the sdk.
I went for the simplest solution, using a header. The system can be abused
of course but there is still a limit which is the most important (no unbounded
number of requests). Worse case scenaio: with this commit we allowed
twice the number of requests.
Both rate-limit buckets have the same size for now. We can make them
have different size in the future if needs arises