fix: allow server to startup when ES isn't responding
Server currently doesn't start if ES migration fails for whatever
reason. Which means that during a ES outage server couldn't be deployed
or scaled up.
This commit ensures the server still starts up even if it cannot connect
to ES (only if ConnectionError, other errors still cause the startup to
fail)
The risk is to have the ES migration to fail silently, the benefits is
server can operate fully when ES is down