fix(webhook-sync-failure): fix false report (#4579)
<!-- Describe the problem and your solution --> 
When there is an auth failure (credentials can't refresh) and a sync
fails the webhook sent is always of status `INITIAL` when it should be
incremental. This updates it so we send the `lastSyncDate` so the
`operation` can be determined correctly

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

<!-- Summary by @propel-code-bot -->

---

This change addresses an issue where, upon authentication failure during
a sync, the webhook payload incorrectly reported the sync status as
'INITIAL' rather than 'incremental'. The update ensures the proper
tracking of sync status by passing the lastSyncDate in the payload so
the operation type is correctly determined downstream.

*This summary was automatically generated by @propel-code-bot*