docs: change date for end user release (#4580)
## Changes

- change date for end user release
Since I wasn't able to merge it last week

<!-- Summary by @propel-code-bot -->

---

This PR updates the developer changelog documentation by changing the
scheduled dates for a previously announced 'end user release'. The main
reason for this modification is that the update could not be merged last
week as originally planned, requiring the public-facing release and
timeline dates to be shifted forward.

*This summary was automatically generated by @propel-code-bot*