fix(jobs): fail quicker when runner is not available
Back in the old days, the requests dispatching tasks to runner were
inlined :O and timeouts were disabled so they could accomodate long running
scripts. It has not been working like this for a while now but we kept
the no timeouts which mean the requests from jobs to runner can take a
long time to fail when runner is overwhelmed/not reachable. (I have seen
2+ mins in traces).
This commit adds timeouts for requests to runner so it fails quicker.
This is one of the reason webhooks couldn't be processed quickly enough
when we experienced some spikes in the past few days. Not that they
would have been executed since the customer runner was not healthy but
they would have failed fast