fix: don't emit task completion events for syncs/webhooks
task completion events for webhooks and syncs are never listened to. Only
actions and onEvents tasks are actually awaited. Which means that the
majority of events are emitted and always discarded
With this commit, task completion events will only be emitted if actions
or onEvents, reducing even more the number of task events